Fix : 맞춤 셀렉션 페치마다 데이터 새로 가져오는 오류 수정
