+++ src/app/page.tsx
@@ -1,13 +1,13 @@
 import BannerSection from "@/components/main/BannerSection";
 import PopularSection from "@/components/main/PopularSection";
-import RecommendationSection from "@/components/main/RecommendationSection";
-import InterestingSection from "@/components/main/InterestingSection";
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
 import { fetchHandler } from "@/http/http";
 import { Suspense } from "react";
 import PageLoading from "@/components/common/PageLoading";
 import PageError from "@/components/common/PageError";
+import RecommendationSection from "@/components/main/RecommendationSection";
+import DisplaySection from "@/components/main/DisplaySection";
 
 export const dynamic = "force-dynamic";
 
@@ -18,9 +18,9 @@ export default async function Home() {
       <Suspense fallback={<PageLoading />}>
         <main className="w-[1086px] bg-grey0 border border-solid border-grey2 m-auto pt-10 flex flex-col gap-10 box-border h-auto">
           <BannerSection />
-          <RecommendationSection />
+          <DisplaySection />
           <PopularSection selections={popularSelections} />
-          <InterestingSection />
+          <RecommendationSection />
         </main>
       </Suspense>
     );
